Silence a few C4244 warnings (including some that were repeated many times):  - add some explicit casts  - disable C4244 warnings for libavutil/common.h
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4712 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/PPageSubStyle.cpp
@@ -170,9 +170,9 @@ BOOL CPPageSubStyle::OnInitDialog()
 	m_scaleyspin.SetRange32(-10000, 10000);
 
 	m_borderstyle = m_stss.borderStyle;
-	m_borderwidth = min(m_stss.outlineWidthX, m_stss.outlineWidthY);
+	m_borderwidth = (int)min(m_stss.outlineWidthX, m_stss.outlineWidthY);
 	m_borderwidthspin.SetRange32(0, 10000);
-	m_shadowdepth = min(m_stss.shadowDepthX, m_stss.shadowDepthY);
+	m_shadowdepth = (int)min(m_stss.shadowDepthX, m_stss.shadowDepthY);
 	m_shadowdepthspin.SetRange32(0, 10000);
 
 	m_screenalignment = m_stss.scrAlignment-1;
